Mas o menos todo hecho
