+++ frontend/src/App.js
@@ -10,20 +10,26 @@ import UserSettings from "./pages/userSettings/UserSettings";
 import SponsorsCRUD from "./pages/Sponsors/SponsorsCRUD";
 import UserCRUD from "./pages/userSettings/userCRUD";
 import RunnersCRUD from "./pages/Runners/RunnerCRUD";
+import RunnersCRUDuser from "./pages/Runners/RunnerCRUDuser";
+import HomeAdmin from "./pages/Home/HomeAdmin";
+import RaceDataAdmin from "./pages/RaceData/RaceDataAdmin";
 function App() {
   return (
     
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Login/>}/>
+        <Route path="/homeadmin" element={<HomeAdmin/>}/>
         <Route path="/home" element={<Home/>}/>
         <Route path="/signin" element={<Signin/>}/>
         <Route path="/login" element={<Login/>}/>
         <Route path="/racedata/:id" element={<RaceData/>}/>
+        <Route path="/racedataAdmin/:id" element={<RaceDataAdmin/>}/>
         <Route path="/runners/:id" element={<RunnersPage/>}/>
         <Route path="/sponsors/:id" element={<SponsorsPage/>}/>
         <Route path="/usersettings" element={<UserSettings/>}/>
         <Route path="/runners" element={<RunnersCRUD/>}/>
+        <Route path="/runnersuser" element={<RunnersCRUDuser/>}/>
         <Route path="/sponsors" element={<SponsorsCRUD/>}/>
         <Route path="/users" element={<UserCRUD/>}/>
       </Routes>
